chore: use controller
