+++ packages/ai-native/src/browser/widget/inline-diff/index.ts
@@ -1,3 +1,3 @@
-export * from './inline-diff.handler';
+export * from './inline-diff.controller';
 export * from './inline-diff-previewer';
 export * from './inline-diff-widget';
