+++ packages/ai-native/src/browser/widget/inline-hint/inline-hint.controller.ts
@@ -1,12 +1,17 @@
-import { Injectable } from '@opensumi/di';
 import { IDisposable } from '@opensumi/ide-core-common';
+import { Disposable } from '@opensumi/ide-core-common';
 import { ICodeEditor } from '@opensumi/ide-monaco';
+import * as monaco from '@opensumi/ide-monaco';
+import { empty } from '@opensumi/ide-utils/lib/strings';
 
+import { AINativeContextKey } from '../../contextkey/ai-native.contextkey.service';
 import { BaseAIMonacoEditorController } from '../../contrib/base';
+import { AICompletionsService } from '../../contrib/inline-completions/service/ai-completions.service';
+import { InlineInputChatService } from '../inline-input/inline-input.service';
+import { InlineInputPreviewDecorationID } from '../internal.type';
 
-import { InlineHintHandler } from './inline-hint.handler';
+import { InlineHintLineDecoration } from './inline-hint-line-widget';
 
-@Injectable()
 export class InlineHintController extends BaseAIMonacoEditorController {
   public static readonly ID = 'editor.contrib.ai.inline.hint';
 
@@ -14,12 +19,97 @@ export class InlineHintController extends BaseAIMonacoEditorController {
     return editor.getContribution<InlineHintController>(InlineHintController.ID);
   }
 
-  private inlineHintHandler: InlineHintHandler;
-  mount(): IDisposable {
-    this.inlineHintHandler = this.injector.get(InlineHintHandler);
-    this.inlineHintHandler.mountEditor(this.monacoEditor);
-    this.inlineHintHandler.load();
+  private get inlineCompletionsService(): AICompletionsService {
+    return this.injector.get(AICompletionsService);
+  }
 
-    return this.inlineHintHandler;
+  private get inlineInputChatService(): InlineInputChatService {
+    return this.injector.get(InlineInputChatService);
+  }
+
+  mount(): IDisposable {
+    return this.registerHintLineFeature(this.monacoEditor);
+  }
+
+  private registerHintLineFeature(monacoEditor: monaco.ICodeEditor): IDisposable {
+    const hintDisposable = new Disposable();
+    const aiNativeContextKey = this.injector.get(AINativeContextKey, [monacoEditor.contextKeyService]);
+
+    const hideHint = () => {
+      hintDisposable.dispose();
+    };
+
+    const showHint = (position: monaco.Position) => {
+      const model = monacoEditor.getModel();
+      if (!model) {
+        return;
+      }
+
+      if (this.inlineCompletionsService.isVisibleCompletion) {
+        return;
+      }
+
+      const content = model.getLineContent(position.lineNumber);
+      const decorations = model.getLineDecorations(position.lineNumber);
+
+      const isEmpty = content?.trim() === empty;
+      const isEmptySelection = monacoEditor.getSelection()?.isEmpty();
+      const hasPreviewDecoration = decorations.some(
+        (dec) => dec.options.description === InlineInputPreviewDecorationID,
+      );
+
+      if (isEmpty && isEmptySelection && !hasPreviewDecoration) {
+        const inlineHintLineDecoration = this.injector.get(InlineHintLineDecoration, [monacoEditor]);
+        inlineHintLineDecoration.show(position);
+        this.inlineInputChatService.setCurrentVisiblePosition(position);
+
+        aiNativeContextKey.inlineHintWidgetIsVisible.set(true);
+
+        hintDisposable.addDispose(
+          inlineHintLineDecoration.onDispose(() => {
+            this.inlineInputChatService.setCurrentVisiblePosition(undefined);
+            aiNativeContextKey.inlineHintWidgetIsVisible.set(false);
+          }),
+        );
+
+        hintDisposable.addDispose(inlineHintLineDecoration);
+      }
+    };
+
+    this.addDispose(
+      monacoEditor.onDidChangeCursorPosition((e: monaco.editor.ICursorPositionChangedEvent) => {
+        hideHint();
+        showHint(e.position);
+      }),
+    );
+
+    this.addDispose(
+      monacoEditor.onDidFocusEditorWidget(() => {
+        const currentPosition = monacoEditor.getPosition();
+
+        if (currentPosition) {
+          hideHint();
+          showHint(currentPosition);
+        }
+      }),
+    );
+
+    this.addDispose(
+      monacoEditor.onDidBlurEditorWidget(() => {
+        hideHint();
+      }),
+    );
+
+    this.addDispose(
+      this.inlineCompletionsService.onVisibleCompletion((v) => {
+        if (v) {
+          hideHint();
+        }
+      }),
+    );
+
+    this.addDispose(hintDisposable);
+
+    return this;
   }
 }
